+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Display error messages from API
 * Update installation instructions
+* Remove 'Leave a Rating' feature
 
 ## 1.0.0 ##